Crate hlua

Reexports

pub use functions_read::LuaFunction;

pub use lua_tables::LuaTable;

Modules

any

functions_read

lua_tables

userdata

Macros

implement_lua_push!

implement_lua_read!

Structs

Function
Opaque type containing a Rust function or closure.

InsideCallback
Opaque type that represents the Lua context when inside a callback.

Lua
Main object of the library.

LuaContext
Opaque type that contains the raw Lua context.

PushGuard
RAII guard for a value pushed on the stack.

Enums

LuaError
Error that can happen when executing Lua code.

Traits

AsLua
Trait for objects that have access to a Lua context. When using a context returned by a AsLua, you are not allowed to modify the stack.

AsMutLua
Trait for objects that have access to a Lua context. You are allowed to modify the stack, but it must be in the same state as it was when you started.

LuaRead
Types that can be obtained from a Lua context.

Push
Types that can be given to a Lua context, for example with lua.set() or as a return value of a function.

Functions

function0
Wraps a type that implements FnMut so that it can be used by hlua.

function1
Wraps a type that implements FnMut so that it can be used by hlua.

function10
Wraps a type that implements FnMut so that it can be used by hlua.

function2
Wraps a type that implements FnMut so that it can be used by hlua.

function3
Wraps a type that implements FnMut so that it can be used by hlua.

function4
Wraps a type that implements FnMut so that it can be used by hlua.

function5
Wraps a type that implements FnMut so that it can be used by hlua.

function6
Wraps a type that implements FnMut so that it can be used by hlua.

function7
Wraps a type that implements FnMut so that it can be used by hlua.

function8
Wraps a type that implements FnMut so that it can be used by hlua.

function9
Wraps a type that implements FnMut so that it can be used by hlua.