Enum hlua::any::AnyLuaValue

pub enum AnyLuaValue {
    LuaString(String),
    LuaNumber(f64),
    LuaBoolean(bool),
    LuaArray(Vec<(AnyLuaValue, AnyLuaValue)>),
    LuaNil,
    LuaOther,
}

Represents any value that can be stored by Lua

Variants
LuaString(String)LuaNumber(f64)LuaBoolean(bool)LuaArray(Vec<(AnyLuaValue, AnyLuaValue)>)LuaNilLuaOther

The "Other" element is (hopefully) temporary and will be replaced by "Function" and "Userdata". A panic! will trigger if you try to push a Other.