Function hlua::function2

pub fn function2<Z, R, A, B>(f: Z) -> Function<Z, (A, B), R> where Z: FnMut(A, B) -> R

Wraps a type that implements FnMut so that it can be used by hlua.

This is only needed because of a limitation in Rust's inferrence system.