Struct hlua::lua_tables::LuaTable

pub struct LuaTable<L> {
    // some fields omitted
}

Represents a table stored in the Lua context.

Loading this type mutably borrows the Lua context.

Methods

impl<L> LuaTable<L> where L: AsMutLua

fn into_inner(self) -> L

Destroys the LuaTable and returns its inner Lua context. Useful when it takes Lua by value.

fn iter<K, V>(&mut self) -> LuaTableIterator<L, K, V>

Iterates over the elements inside the table.

fn get<'a, R, I>(&'a mut self, index: I) -> Option<R> where R: LuaRead<PushGuard<&'a mut LuaTable<L>>>, I: for<'b> Push<&'b mut &'a mut LuaTable<L>>

Loads a value in the table given its index.

fn into_get<'a, R, I>(self, index: I) -> Result<R, PushGuard<Self>> where R: LuaRead<PushGuard<LuaTable<L>>>, I: for<'b> Push<&'b mut LuaTable<L>>

Loads a value in the table, with the result capturing the table by value.

fn set<'s, I, V>(&'s mut self, index: I, value: V) where I: for<'a> Push<&'a mut &'s mut LuaTable<L>>, V: for<'a> Push<&'a mut &'s mut LuaTable<L>>

Inserts or modifies an elements of the table.

fn empty_array<'s, I>(&'s mut self, index: I) -> LuaTable<PushGuard<&'s mut LuaTable<L>>> where I: for<'a> Push<&'a mut &'s mut LuaTable<L>> + Clone

Inserts an empty array, then loads it.

fn get_or_create_metatable(self) -> LuaTable<PushGuard<L>>

Obtains or create the metatable of the table.