Struct hlua::InsideCallback

pub struct InsideCallback {
    // some fields omitted
}

Opaque type that represents the Lua context when inside a callback.

Some types (like Result) can only be returned from a callback and not written inside a Lua variable. This type is here to enforce this restriction.