Trait hlua::AsLua

pub unsafe trait AsLua {
    fn as_lua(&self) -> LuaContext;
}

Trait for objects that have access to a Lua context. When using a context returned by a AsLua, you are not allowed to modify the stack.

Required Methods

fn as_lua(&self) -> LuaContext