Trait hlua::Push

pub trait Push<L> where L: AsMutLua {
    fn push_to_lua(self, lua: L) -> PushGuard<L>;
}

Types that can be given to a Lua context, for example with lua.set() or as a return value of a function.

Required Methods

fn push_to_lua(self, lua: L) -> PushGuard<L>

Pushes the value on the top of the stack.

Must return a guard representing the elements that have been pushed.

You can implement this for any type you want by redirecting to call to another implementation (for example 5.push_to_lua) or by calling userdata::push_userdata.